+++ packages/vexide-core/src/time.rs
@@ -2,10 +2,28 @@
 
 use core::time::Duration;
 
-use vex_sdk::vexSystemPowerupTimeGet;
+use vex_sdk::{vexSystemHighResTimeGet, vexSystemPowerupTimeGet};
 
-/// Returns the duration that the brain has been powered on for.
+/// Returns the duration that the brain has been powered on.
+///
+/// # Precision
+///
+/// The returned [`Duration`] has a precision of 1 microsecond.
 #[must_use]
-pub fn uptime() -> Duration {
+pub fn system_uptime() -> Duration {
     Duration::from_micros(unsafe { vexSystemPowerupTimeGet() })
+}
+
+/// Returns the duration that the brain's user processor has been
+/// running.
+///
+/// This is effectively the amount of time that the current user
+/// program has been running.
+///
+/// # Precision
+///
+/// The returned [`Duration`] has a precision of 1 microsecond.
+#[must_use]
+pub fn user_uptime() -> Duration {
+    Duration::from_micros(unsafe { vexSystemHighResTimeGet() })
 }